Commit Update 1.1

+++ ESP32_GuitarCable/include/FIFO.h
@@ -10,7 +10,7 @@
 #define PREBUFF_LEN 512
 
 /* External Variables */
-extern uint8_t Task_buff[PREBUFF_LEN];
+extern uint16_t Task_buff[PREBUFF_LEN];
 extern volatile uint16_t head;
 extern volatile uint16_t tail;
 
@@ -23,10 +23,10 @@ uint8_t isFull(void);
 uint8_t isEmpty(void);
 
 // Push Sample into FIFO Queue
-uint8_t push_Sample(uint8_t sample);
+uint8_t push_Sample(uint16_t sample);
 
 // Pop Sample out of FIFO Queue
-uint8_t pop_Sample(uint8_t *sample);
+uint8_t pop_Sample(uint16_t *sample);
 
 // Check how Many Samples are Pre-Buffered
 uint16_t sampleCount();
