+++ ESP32_GuitarCable/src/LPF.cpp
@@ -3,13 +3,13 @@
 
 
 /* Global Variables */
-uint8_t fir_buffer[FIR_TAPS] = {0};
-uint8_t fir_idx = 0;
+uint16_t fir_buffer[FIR_TAPS] = {0};
+uint16_t fir_idx = 0;
 
 
 /* Functions */
 // Create Multi-Tap FIR Filter
-uint8_t apply_fir_filter(uint8_t new_sample) {
+uint16_t apply_fir_filter(uint16_t new_sample) {
     fir_buffer[fir_idx] = new_sample;
     fir_idx = (fir_idx + 1) % FIR_TAPS;
 
@@ -18,5 +18,5 @@ uint8_t apply_fir_filter(uint8_t new_sample) {
         sum += fir_buffer[i];
     }
 
-    return (uint8_t)(sum / FIR_TAPS);
+    return (uint16_t)(sum / FIR_TAPS);
 }
